+++ feature/login/src/main/java/com/twix/login/LoginViewModel.kt
@@ -7,9 +7,9 @@ import com.twix.domain.login.LoginResult
 import com.twix.domain.model.OnboardingStatus
 import com.twix.domain.repository.AuthRepository
 import com.twix.domain.repository.OnBoardingRepository
-import com.twix.login.model.LoginIntent
-import com.twix.login.model.LoginSideEffect
-import com.twix.login.model.LoginUiState
+import com.twix.login.contract.LoginIntent
+import com.twix.login.contract.LoginSideEffect
+import com.twix.login.contract.LoginUiState
 import com.twix.ui.base.BaseViewModel
 import kotlinx.coroutines.launch
