+++ feature/login/src/main/java/com/twix/login/contract/LoginSideEffect.kt
@@ -1,4 +1,4 @@
-package com.twix.login.model
+package com.twix.login.contract
 
 import com.twix.designsystem.components.toast.model.ToastType
 import com.twix.domain.model.OnboardingStatus
